Rename cssStatementHasEmptyBlock to hasEmptyBlock
