+++ src/rules/block-opening-brace-newline-before/index.js
@@ -1,6 +1,6 @@
 import {
   hasBlock,
-  cssStatementHasEmptyBlock,
+  hasEmptyBlock,
   blockString,
   cssStatementStringBeforeBlock,
   report,
@@ -42,7 +42,7 @@ export default function (expectation) {
     function check(statement) {
 
       // Return early if blockless or has an empty block
-      if (!hasBlock(statement) || cssStatementHasEmptyBlock(statement)) { return }
+      if (!hasBlock(statement) || hasEmptyBlock(statement)) { return }
 
       const beforeBrace = cssStatementStringBeforeBlock(statement)
