+++ src/rules/block-opening-brace-space-after/index.js
@@ -1,7 +1,7 @@
 import {
   blockString,
   hasBlock,
-  cssStatementHasEmptyBlock,
+  hasEmptyBlock,
   cssStatementStringBeforeBlock,
   report,
   ruleMessages,
@@ -42,7 +42,7 @@ export default function (expectation) {
 
     function check(statement) {
       // Return early if blockless or has an empty block
-      if (!hasBlock(statement) || cssStatementHasEmptyBlock(statement)) { return }
+      if (!hasBlock(statement) || hasEmptyBlock(statement)) { return }
 
       checker.after({
         source: blockString(statement),
