+++ src/utils/hasEmptyBlock.js
