+++ src/utils/index.js
@@ -5,7 +5,7 @@ export { default as functionArgumentsSearch } from "./functionArgumentsSearch"
 export { default as cssNodeContextLookup } from "./cssNodeContextLookup"
 export { default as blockString } from "./blockString"
 export { default as hasBlock } from "./hasBlock"
-export { default as cssStatementHasEmptyBlock } from "./cssStatementHasEmptyBlock"
+export { default as hasEmptyBlock } from "./hasEmptyBlock"
 export { default as cssStatementStringBeforeBlock } from "./cssStatementStringBeforeBlock"
 export { default as cssWordIsVariable } from "./cssWordIsVariable"
 export { default as declarationValueIndexOffset } from "./declarationValueIndexOffset"
